Model TDD: REFACTOR int to NSUinteger

+++ CS345 Pente/CS345 Pente/CSPGollumInTheCloset.h
@@ -10,7 +10,7 @@
 #import "CSPPenteGameModelInterface.h"
 
 @interface CSPGollumInTheCloset : NSObject <CSPPenteGameModelInterface>
-@property (assign) int turnNumber;
+@property (assign) NSUInteger turnNumber;
 @property (assign) CSPGameOverID gameOverState;
 
 @end
